Use the same rows in the table and chart
