+++ src/components/HrTools/GoalCalculator/SummaryReport/Steps/PresentingYourGoalStep/PresentingYourGoal.tsx
@@ -1,20 +1,10 @@
 import React, { useMemo } from 'react';
-import { Box, styled, useMediaQuery } from '@mui/material';
+import { styled } from '@mui/material';
 import { useTranslation } from 'react-i18next';
-import {
-  Cell,
-  Legend,
-  Pie,
-  PieChart,
-  ResponsiveContainer,
-  Tooltip,
-} from 'recharts';
 import { MonthlyNeedsCard } from 'src/components/HrTools/Shared/GoalPresentation/MonthlyNeedsCard';
 import { PersonalInfoCard } from 'src/components/HrTools/Shared/GoalPresentation/PersonalInfoCard';
 import { PresentationCard } from 'src/components/HrTools/Shared/GoalPresentation/PresentationCard';
-import { useLocale } from 'src/hooks/useLocale';
-import { currencyFormat, percentageFormat } from 'src/lib/intlFormat';
-import theme from 'src/theme';
+import { SupportNeedsChart } from 'src/components/HrTools/Shared/GoalPresentation/SupportNeedsChart';
 import { useGoalCalculator } from '../../../Shared/GoalCalculatorContext';
 import { hasStaffSpouse } from '../../../Shared/calculateTotals';
 
@@ -29,37 +19,6 @@ const PrintableContent = styled('div')(({ theme }) => ({
   },
 }));
 
-const ChartContainer = styled(Box)({
-  '@media print': {
-    width: '100% !important',
-    height: '350px !important',
-    '.recharts-wrapper': {
-      height: '100px !important',
-      width: '100% !important',
-    },
-    '.recharts-surface': {
-      height: '350px !important',
-      width: '100% !important',
-    },
-    '.recharts-legend-wrapper': {
-      fontSize: '16px !important',
-      paddingRight: '20px !important',
-      textAlign: 'center !important',
-      top: '100px !important',
-    },
-  },
-
-  '.recharts-legend-item .recharts-surface': {
-    width: '16px !important',
-    height: '16px !important',
-    top: '2px',
-  },
-
-  '.recharts-legend-item text': {
-    dominantBaseline: 'middle',
-  },
-});
-
 interface PresentingYourGoalProps {
   supportRaised: number;
 }
@@ -68,16 +27,16 @@ export const PresentingYourGoal: React.FC<PresentingYourGoalProps> = ({
   supportRaised,
 }) => {
   const { t } = useTranslation();
-  const locale = useLocale();
-  const isMobile = useMediaQuery(theme.breakpoints.down('lg'));
   const {
     goalCalculationResult: { data },
     goalTotals,
   } = useGoalCalculator();
   const goalCalculation = data?.goalCalculation;
+  const married = hasStaffSpouse(goalCalculation?.familySize);
 
   const supportNeeds = useMemo(
     () => ({
+      married,
       salary: goalTotals.netMonthlySalary,
       ministryExpenses: goalTotals.ministryExpensesTotal + goalTotals.attrition,
       benefits: goalTotals.benefitsCharge,
@@ -87,107 +46,25 @@ export const PresentingYourGoal: React.FC<PresentingYourGoalProps> = ({
       adminCharge:
         goalTotals.overallSubtotalWithAdmin - goalTotals.overallSubtotal,
     }),
-    [goalTotals],
+    [married, goalTotals],
   );
-
-  const presentationData = useMemo(
-    () => [
-      { name: 'Salary', value: supportNeeds.salary },
-      { name: 'Ministry Expenses', value: supportNeeds.ministryExpenses },
-      { name: 'Benefits', value: supportNeeds.benefits },
-      {
-        name: 'Social Security and Taxes',
-        value: supportNeeds.socialSecurityAndTaxes,
-      },
-      {
-        name: 'Voluntary 403b Retirement Plan',
-        value: supportNeeds.voluntaryRetirement,
-      },
-      {
-        name: 'Administrative Charge',
-        value: supportNeeds.adminCharge,
-      },
-    ],
-    [supportNeeds],
-  );
-
-  const total = useMemo(
-    () => presentationData.reduce((sum, entry) => sum + entry.value, 0),
-    [presentationData],
-  );
-
-  // Consider adding more brand colors to theme.
-  const chartColors = [
-    theme.palette.primary.main,
-    theme.palette.secondary.main,
-    theme.palette.success.main,
-    theme.palette.warning.main,
-    theme.palette.error.main,
-    theme.palette.info.main,
-  ];
 
   return (
     <PrintableContent>
       <PersonalInfoCard
         firstName={goalCalculation?.firstName ?? ''}
-        spouseFirstName={
-          hasStaffSpouse(goalCalculation?.familySize)
-            ? goalCalculation?.spouseFirstName
-            : null
-        }
+        spouseFirstName={married ? goalCalculation?.spouseFirstName : null}
         lastName={goalCalculation?.lastName ?? ''}
         ministryLocation={goalCalculation?.ministryLocation ?? undefined}
       />
 
       <MonthlyNeedsCard
-        {...supportNeeds}
-        married={hasStaffSpouse(goalCalculation?.familySize)}
+        monthlyNeeds={supportNeeds}
         supportRaised={supportRaised}
       />
 
       <PresentationCard title={t('Monthly Support Breakdown')}>
-        <ChartContainer height={500}>
-          <ResponsiveContainer width="100%">
-            <PieChart>
-              <Pie
-                data={presentationData}
-                dataKey="value"
-                nameKey="name"
-                cx="50%"
-                cy="50%"
-                outerRadius={isMobile ? 130 : 180}
-                cornerRadius={theme.shape.borderRadius}
-              >
-                {presentationData.map((entry, index) => (
-                  <Cell
-                    key={entry.name}
-                    fill={chartColors[index % chartColors.length]}
-                  />
-                ))}
-              </Pie>
-              <Tooltip
-                formatter={(value) => {
-                  return `${currencyFormat(
-                    Number(value),
-                    'USD',
-                    locale,
-                  )} (${percentageFormat(Number(value) / total, locale)})`;
-                }}
-              />
-              <Legend
-                layout="vertical"
-                align={isMobile ? 'center' : 'right'}
-                verticalAlign={isMobile ? 'bottom' : 'middle'}
-                wrapperStyle={{
-                  fontSize: isMobile
-                    ? theme.typography.subtitle1.fontSize
-                    : theme.typography.h5.fontSize,
-                  maxWidth: isMobile ? 600 : 300,
-                }}
-              />
-            </PieChart>
-          </ResponsiveContainer>
-        </ChartContainer>
+        <SupportNeedsChart monthlyNeeds={supportNeeds} />
       </PresentationCard>
     </PrintableContent>
   );
